requirements.txt
openai~=0.27.2
pyttsx3~=2.90
requests~=2.28.2
pillow~=9.4.0